+++ tests/test_events.py
@@ -5,7 +5,7 @@ import werobot.testing
 
 def test_events():
     robot = werobot.WeRoBot(token=werobot.utils.generate_token())
-    t = [False, '']
+    t = [False, '', 'NOTCHANGED']
 
     @robot.subscribe
     def first(message):
@@ -22,7 +22,7 @@ def test_events():
         <Event><![CDATA[subscribe]]></Event>
     </xml>
     """
-    assert tester.send_xml(xml) == 'Hi'
+    assert tester.send_xml(xml) == 'Hi', tester.send_xml(xml)
 
     @robot.unsubscribe
     def second(message):
@@ -56,3 +56,42 @@ def test_events():
     """
     tester.send_xml(xml)
     assert t[1] == 'EVENTKEY'
+
+    @robot.key_click('MYEVENT')
+    def key(message):
+        return message.key
+
+    @robot.key_click('NOTMYEVENT')
+    def nokey(message):
+        t[2] = 'CHANGED'
+        return message.key
+
+    xml = """
+    <xml>
+        <ToUserName><![CDATA[toUser]]></ToUserName>
+        <FromUserName><![CDATA[fromUser]]></FromUserName>
+        <CreateTime>123456789</CreateTime>
+        <MsgType><![CDATA[event]]></MsgType>
+        <Event><![CDATA[CLICK]]></Event>
+        <EventKey><![CDATA[MYEVENT]]></EventKey>
+    </xml>
+    """
+
+    assert tester.send_xml(xml) == 'MYEVENT'
+    assert t[2] == 'NOTCHANGED'
+
+    @robot.key_click('ARGS')
+    def key():
+        return 'ARGS'
+
+    xml = """
+    <xml>
+        <ToUserName><![CDATA[toUser]]></ToUserName>
+        <FromUserName><![CDATA[fromUser]]></FromUserName>
+        <CreateTime>123456789</CreateTime>
+        <MsgType><![CDATA[event]]></MsgType>
+        <Event><![CDATA[CLICK]]></Event>
+        <EventKey><![CDATA[ARGS]]></EventKey>
+    </xml>
+    """
+    assert tester.send_xml(xml) == 'ARGS'