+++ werobot/__init__.py
@@ -1,8 +1,5 @@
-__version__ = '0.4.1'
+__version__ = '0.5.0'
 __author__ = 'whtsky'
 __license__ = 'MIT'
 
-try:
-    from .robot import WeRoBot
-except:
-    pass
+from werobot.robot import WeRoBot