+++ werobot/session/filestorage.py
@@ -6,7 +6,7 @@ try:
 except ImportError:
     import dbm
 
-from . import SessionStorage
+from werobot.session import SessionStorage
 from werobot.utils import json_loads, json_dumps